figuring out how to maintain organization
